Start to implement a cargo test integration
The general idea is to use the tooling from cargo directly. Cargo is
available as a library, so it is possible to call into it. This allows
to run the tests directly using Cargo. Note that this results in
compiling Rust code from within BEAM, so a dirty scheduler might have to be used.